changed x_coord from int to float in drive gesture recognition scripts

+++ Assets/Scripts/Custom_Gestures/DrivePitchGesture.cs
@@ -38,7 +38,7 @@ public class DrivePitchGesture : MonoBehaviour
 
 
 	[Range (0, 10)]
-	public int x_movement = 1;
+	public float x_movement = 1;
 
 	public HandController hc;
